fix: изменено взаимодействие игр и участников
- сущность Participant и ее прилагающиеся - удалены;
- в сущность игры добавлен список участников;
- GameDto теперь выводит id пользователей, а не всю сущность;
- в балансе игры теперь допускаюся копейки;
- убраны лишние логи
- в GameMapper добавлена возможность обновлять сущность, а также конвертировать ее список участников в список id участников;
- добавлен 1 тест в постман коллекцию.

+++ src/main/java/ru/karyeragame/paymentsystem/game/dto/NewGameDto.java
@@ -19,5 +19,5 @@ public class NewGameDto {
     private Long createdBy;
     @NotNull
     @PositiveOrZero
-    private Integer startBalance;
+    private Float startBalance;
 }
